+++ protocols/request-response/src/lib.rs
@@ -599,7 +599,7 @@ where
         new: &ConnectedPoint,
     ) {
         let new_address = match new {
-            ConnectedPoint::Dialer { address } => Some(address.clone()),
+            ConnectedPoint::Dialer { address, .. } => Some(address.clone()),
             ConnectedPoint::Listener { .. } => None,
         };
         let connections = self
@@ -631,7 +631,7 @@ where
         _errors: Option<&Vec<Multiaddr>>,
     ) {
         let address = match endpoint {
-            ConnectedPoint::Dialer { address } => Some(address.clone()),
+            ConnectedPoint::Dialer { address, .. } => Some(address.clone()),
             ConnectedPoint::Listener { .. } => None,
         };
         self.connected